fix: resolve foreign key data type conflicts and stabilize server

+++ models/Report.js
@@ -3,7 +3,7 @@ const sequelize = require('../config/database');
 
 const Report = sequelize.define('Report', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  user_id: { type: DataTypes.BIGINT, allowNull: false },
+  user_id: { type: DataTypes.INTEGER, allowNull: false },
   category: { 
     type: DataTypes.ENUM('accident', 'new checkpoint', 'weather', 'closure'), 
     allowNull: false
