+++ models/StatusHistory.js
@@ -3,9 +3,9 @@ const sequelize = require('../config/database');
 
 const StatusHistory = sequelize.define('StatusHistory', {
   id: { type: DataTypes.INTEGER, primaryKey: true, autoIncrement: true },
-  checkpoint_id: { type: DataTypes.BIGINT, allowNull: false },
+  checkpoint_id: { type: DataTypes.INTEGER, allowNull: false },
   status: { type: DataTypes.STRING(45), allowNull: false },
-  changed_by: { type: DataTypes.BIGINT, allowNull: false },
+  changed_by: { type: DataTypes.INTEGER, allowNull: false },
   reason: { type: DataTypes.TEXT },
   timestamp: { type: DataTypes.DATE, defaultValue: DataTypes.NOW }
 }, {
